requirements.txt
Flask==2.3.2
gunicorn==20.1.0
prometheus-client==0.16.0
pyyaml==6.0
requests==2.28.2
